+++ Src/OSF/nginx/src/core/ngx_crypt.c
@@ -143,7 +143,7 @@ static ngx_int_t ngx_crypt_ssha(ngx_pool_t * pool, u_char * key, u_char * salt, 
 	/* decode base64 salt to find out true salt */
 	encoded.data = salt + sizeof("{SSHA}") - 1;
 	encoded.len = ngx_strlen(encoded.data);
-	len = MAX(ngx_base64_decoded_length(encoded.len), 20);
+	len = MAX(ngx_base64_decoded_length(encoded.len), static_cast<size_t>(20));
 	decoded.data = static_cast<u_char *>(ngx_pnalloc(pool, len));
 	if(decoded.data == NULL) {
 		return NGX_ERROR;
